broken: saving flashcards to persistent storage
